+++ lib/details/types/audios_from_type.dart
@@ -1,7 +1,13 @@
 part of on_audio_query;
 
+/// Defines where audios will be acquired.
 enum AudiosFromType {
+  /// Audios from specific Album.
   ALBUM,
+
+  /// Audios from specific Artist.
   ARTIST,
+
+  /// Audios from specific Playlist.
   PLAYLIST,
 }